requirements.txt
## The following requirements were added by pip freeze:
certifi==2020.6.20
cycler==0.10.0
iso8601==0.1.13
kiwisolver==1.2.0
matplotlib==3.3.2
numpy==1.19.2
pandas==1.1.2
Pillow==7.2.0
pyparsing==2.4.7
PyQt5==5.15.1
PyQt5-sip==12.8.1
pyserial==3.4
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
scipy==1.5.2
six==1.15.0
